backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.2
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-sso==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
oauthlib==3.2.2
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.2
